requirements.txt
netmiko
paho-mqtt